+++ source/gui/panconvert_diag_toformat.py
@@ -2,7 +2,7 @@
 
 # Form implementation generated from reading ui file 'panconvert_diag_toformat.ui'
 #
-# Created: Tue Jan 17 09:00:06 2017
+# Created: Sun Feb 19 15:21:25 2017
 #      by: PyQt5 UI code generator 5.3.2
 #
 # WARNING! All changes made in this file will be lost!
@@ -12,25 +12,20 @@ from PyQt5 import QtCore, QtGui, QtWidgets
 class Ui_To_Format_Dialog(object):
     def setupUi(self, To_Format_Dialog):
         To_Format_Dialog.setObjectName("To_Format_Dialog")
-        To_Format_Dialog.resize(274, 402)
-        self.gridLayout_2 = QtWidgets.QGridLayout(To_Format_Dialog)
-        self.gridLayout_2.setObjectName("gridLayout_2")
-        self.gridLayout = QtWidgets.QGridLayout()
+        To_Format_Dialog.resize(224, 234)
+        self.gridLayout = QtWidgets.QGridLayout(To_Format_Dialog)
         self.gridLayout.setObjectName("gridLayout")
-        self.ButtonCancel = QtWidgets.QPushButton(To_Format_Dialog)
-        self.ButtonCancel.setObjectName("ButtonCancel")
-        self.gridLayout.addWidget(self.ButtonCancel, 0, 0, 1, 1)
-        self.ButtonInfo = QtWidgets.QPushButton(To_Format_Dialog)
-        self.ButtonInfo.setObjectName("ButtonInfo")
-        self.gridLayout.addWidget(self.ButtonInfo, 0, 2, 1, 1)
-        self.ButtonMoreInfo = QtWidgets.QPushButton(To_Format_Dialog)
-        self.ButtonMoreInfo.setObjectName("ButtonMoreInfo")
-        self.gridLayout.addWidget(self.ButtonMoreInfo, 0, 1, 1, 1)
-        self.gridLayout_2.addLayout(self.gridLayout, 2, 0, 1, 1)
         self.textBrowser = QtWebKitWidgets.QWebView(To_Format_Dialog)
+        self.textBrowser.setMinimumSize(QtCore.QSize(200, 100))
         self.textBrowser.setUrl(QtCore.QUrl("about:blank"))
         self.textBrowser.setObjectName("textBrowser")
-        self.gridLayout_2.addWidget(self.textBrowser, 1, 0, 1, 1)
+        self.gridLayout.addWidget(self.textBrowser, 0, 0, 1, 1)
+        self.verticalLayout = QtWidgets.QVBoxLayout()
+        self.verticalLayout.setObjectName("verticalLayout")
+        self.ButtonCancel = QtWidgets.QPushButton(To_Format_Dialog)
+        self.ButtonCancel.setObjectName("ButtonCancel")
+        self.verticalLayout.addWidget(self.ButtonCancel)
+        self.gridLayout.addLayout(self.verticalLayout, 1, 0, 1, 1)
 
         self.retranslateUi(To_Format_Dialog)
         QtCore.QMetaObject.connectSlotsByName(To_Format_Dialog)
@@ -39,7 +34,5 @@ class Ui_To_Format_Dialog(object):
         _translate = QtCore.QCoreApplication.translate
         To_Format_Dialog.setWindowTitle(_translate("To_Format_Dialog", "Pandoc To Formats"))
         self.ButtonCancel.setText(_translate("To_Format_Dialog", "Cancel"))
-        self.ButtonInfo.setText(_translate("To_Format_Dialog", "Formats"))
-        self.ButtonMoreInfo.setText(_translate("To_Format_Dialog", "Help"))
 
 from PyQt5 import QtWebKitWidgets